Increase color array

+++ saco/game/util.h
@@ -25,7 +25,7 @@ void DecompressNormalVector(VECTOR * vec, C_VECTOR1 * c1);
 float FloatDifference(float f1, float f2);
 float FloatOffset(float f1, float f2);
 
-void __stdcall SetRadarColor(BYTE nIndex,DWORD dwColor);
+void __stdcall SetRadarColor(int nIndex,DWORD dwColor);
 
 void __stdcall WorldRemoveEntity(DWORD *dwEnt);
 void __stdcall WorldAddEntity(DWORD *dwEnt);
